+++ ntex-net/CHANGES.md
@@ -1,6 +1,8 @@
 # Changes
 
-* Add windows IOCP driver
+## [3.14.0] - 2026-08-12
+
+* Add windows IOCP driver (neon-iocp)
 
 ## [3.13.1] - 2026-07-17